L2 and BCMP use the new adin driver (#37)
* feat(network): assume 1 device with 2 ports

This commit is WIP. Two L2 tests are still failing,
and I haven't ported the TX/RX functionality yet.

* fix(l2): must start with BmOK in bm_l2_netif_set_power

Because the error value started at BmEINVAL,
we were never actually calling the functions within the
bm_err_check macros.

This commit also reduces a level of indirection.
L2 now uses the NetworkDevice struct directly
rather than a pointer to it.
This cleans up the code significantly
and enabled me to find the bm_err_check bug.

* fix(l2): remove unused code

* fix: unify adin driver mocking

The adin driver test needs to exercise internal driver code,
but the L2 test needs a stub NetworkDevice.

Add include guards to bm_adin2111.h which is required for the L2 unit
test suite, since it needs to include both l2.h and mock_bm_adin2111.h,
both of which need to include bm_adin2111.h.

* fix: move callbaks from Adin2111 to NetworkDevice

This allows L2 to set the receive callback

* refactor: we no longer need the Adin2111 struct

* set network device link change callback in bcmp

* fix: rethinking where we require a pointer

Before this it was possible to set a callback only in one copy of the
NetworkDevice, without it actually being visible to the driver.

* fix: double definition of fff

* test: demonstrate setting callback before adin init

* refactor: remove adin2111 macro from l2 and bcmp

L2 and BCMP should only deal with a NetworkDevice and have no knowledge
of the underlying ADIN2111. This commit adds a num_ports function to the
NetworkDeviceTrait and calls it to set the number of ports in init for
L2 and topology.